Added completed PIA list view
