+++ src/frontend/src/components/common/Navbar/navPages.ts
@@ -7,6 +7,12 @@ export const NavPages = [
   },
   {
     id: 2,
+    label: 'Completed PIAs',
+    link: '/pia/completed',
+    enable: true,
+  },
+  {
+    id: 3,
     label: 'Create new',
     link: '/ppq',
     enable: true,
