+++ src/frontend/src/components/public/ListViewRender/tableProperties.ts
