+++ src/frontend/src/components/public/PIAListTable/interface.ts
@@ -1,4 +1,4 @@
-import { TableHeadingPropertiesType } from '../../../pages/PIAListPage/tableProperties';
+import { TableHeadingPropertiesType } from '../ListViewRender/tableProperties';
 import { IPiaForm } from '../../../types/interfaces/pia-form.interface';
 
 export interface IDataTable {
