+++ src/frontend/src/pages/PIACompletedList/index.tsx
@@ -1,0 +1,7 @@
+import ListViewRender from '../../components/public/ListViewRender';
+
+const PIACompletedList = () => {
+  return <ListViewRender showCompleted={true} />;
+};
+
+export default PIACompletedList;
