+++ src/frontend/src/routes/router.tsx
@@ -9,6 +9,7 @@ import LandingPage from '../pages/LandingPage/LandingPage';
 import PIAIntakeFormPage from '../pages/PIAForm';
 import PIAIntakeResultsPage from '../pages/PIAIntakeResultsPage';
 import PIAList from '../pages/PIAListPage';
+import PIACompletedList from '../pages/PIACompletedList';
 import PPQConnectPage from '../pages/PPQConnectPage';
 import PPQFormPage from '../pages/PPQFormPage/PPQFormPage';
 import PPQLandingPage from '../pages/PPQPage/PPQPage';
@@ -67,6 +68,7 @@ const Router = () => {
           <Route index element={<Navigate relative="path" to="list" />} />
 
           <Route path="list" element={<PIAList />} />
+          <Route path="completed" element={<PIACompletedList />} />
           <Route path="result" element={<PIAIntakeResultsPage />} />
 
           <Route path="new" element={<PIAIntakeFormPage />}>
